Time: 5 ms (72.18%), Space: 44 MB (72.24%) - LeetHub

+++ 0009-palindrome-number/0009-palindrome-number.java
@@ -7,7 +7,7 @@ class Solution {
         long rev=0;
         int temp=m;
         while(temp !=0){
-            int digit=(int )(temp %10);
+            int digit=(temp %10);
             rev=rev*10+digit;
             temp=temp/10;
         }
